// src/app/models/victim.ts
export interface Victim {
    id: number;
    name: string;
    surname: string;
    email: string;
    phoneNumber: string;
    numberOfPeople: number;
    address: string;
    detailedAddress: string;
    urgency: string;
    statusInformation: string;
    mapsAddress: string;
    victimType: string;
}